python_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.1.0
websocket-server==0.6.4